Moving BaseKtorClient into core.data.ktor.
