--- core/data/README.md
@@ -1,7 +1,0 @@
-# Core Data
-
-## Overview
-
-<!-- Explanation of this module's responsibilities. -->
-
-Contains any shared data layer classes, such as our [BaseKtorClient](src/commonMain/kotlin/com/adammcneilly/spacenerd/core/data/BaseKtorClient.kt).
