+++ core/data/ktor/README.md
@@ -1,0 +1,7 @@
+# Core Data
+
+## Overview
+
+<!-- Explanation of this module's responsibilities. -->
+
+Contains any shared data layer classes, such as our [BaseKtorClient](src/commonMain/kotlin/com/adammcneilly/spacenerd/core/data/ktor/BaseKtorClient.kt).
